Fix #34: Use idiomatic config and cache folders
This is idiomatic for every operating system, and gives us better
defaults than just cramming everything under `.bloop`.

For linux systems, for example, the directories are:

1. `~/.cache`
1. `~/.config`
1. `~/.local/share`

Finally, this commit also makes the handling of cache directories more
consistent, making sure that they always exist and that they are
directories before returning them to the users.