Fix #50: Fix plugin switch command implementation
This should probably be reported and fixed upstream.

The implementation of `pluginSwitch` is wrong and does not correctly
reapply the setting when there are other modifications in the state,
produced in the build pipeline.

This commit hot swaps the implementation by the correct one, which uses
the underlying mechanism that the `set` command uses.

Another possible implementation would have been similar to the one we
already have in our default `onLoad`, but this one works so I haven't
tried the other one.