Don't crash with empty `allJars` in ScalaInstance
When creating a `ScalaInstance`, we check that the JARs that we're given
all exist. However, we don't handle the case where the array may be
empty. In this case, they all exist, but we'll create a `ScalaInstance`
with no JARs, which will crash the compiler down the road.

This commit fixes this issue by making us resolve the `ScalaInstance` if
we're given no JARs.