Remove L4J plugin, configure with properties file
I'm having problems when doing a fat JAR with the plugin that we were
using to configure the logger. I preferred that approach at first
because I don't like magic properties files, but looking at the plugin,
I don't think it was buying us much.

The properties file mimics the configuration that we had.