Rework and improve internals
Our previous internal model was okay for a prototype but I don't think
it could stand the test of time.

This commit creates a better internal model that rests on the concept of
`State`. `State` is immutable and contains all the pieces necessary for
compiling and testing projects.

This model has the advantage that is easy to reason about; it isolates
the previous compilation results from the project, it removes the
confusing `Map[String, Project]`, and it allows for more code reuse and
value caching.

There is one `State` per `AbsolutePath`, meaning that by design we don't
support concurrent compilation of the same project. We don't enforce
this yet, though.

Aside the `State` abstraction, this PR also introduces `Dag`, which
represents a direct acyclic graph that is useful when traversing
dependencies of projects and doing reachability analysis.

As a plus, this commit removes all warnings in our codebase.