Add `commonOptions` to state and docs
`CommonOptions` is required in the state so that we can modify the
logger every time the common options change. This can happen when users
pass in command redirecting input and output, or when bloop is run by a
nailgun server.

This is also extremely important for testing. The reconfiguration of
logging is not done in this PR, but will be done in a follow up.