Update loggers with new common options
Every time we change the common options in `State`, via an action, the
loggers have to update their backings so that logs are redirected to
those streams.

This is especially important because of Nailgun. Until this PR, we were
changing the common options in the `Cli` entrypoint but we were not
updating the loggers with those options which meant that we were subject
to losing logs.

I think this fixes the problem that @Duhemm has been experiencing
several times with Nailgun, but on top of that it allows users to
redirect output streams whenever they want (though it's a feature that
should not be used very frequently).

As a plus, we get a nice way to test bloop: we can always redirect
output streams at the test call site and see what the complete output of
it is.