Allow to specify where `BloopLogger` should write
This commit makes us define the streams that should be used by an
instance of `BloopLogger` when constructing the logger. As a result, we
have a clearer relation between each CLI invocation and the assigned
logger, making it easier to have output be written to the correct
stream.

It also solves issues that we had with repeated output (output would be
written to the console as many times as we had loggers created).