Fix logging with Nailgun
Using Nailgun, we re-use loggers with the same name. This means that
between sessions, the old appenders will continue to exist for that
logger. Finally, Log4J will not add the appenders that we create if
appenders with the same name already exist for a given logger.

As a result, we were not adding the right appenders when re-using an
existing logger. This commit fixes that by removing all the current
appenders when creating a new instance of `BloopLogger`.