Replace `discoveredSbtPlugins` in sbt 0.13
This task always triggers compilation, because it doesn't rely on a
dynamic task as it should. This commit replaces the implementation of
this task in sbt-bloop by the same implementation that is used in sbt
1.0 (which is correct and uses a dynamic task).

Because this task is used in a resource generator that is present by
default, we would always trigger compilation when inspecting the
generated resources. This is no longer the case (except if the project
is an sbt plugin).