Add support for resources in `run` and `test`
The sbt plugin is adapted to run `copyResources` when generating the
configuration, which will copy the resources to the classes directory
(thus making them available when running and testing).

Our scripted setup is adapted to copy the generated sources and the
resources out of the scripted environment (scripted tests are run inside
a temporary directory).

Fixes #134