Use Nailgun's `ThreadLocalPrintStream`
It is doing the same thing as our implementation in
`MultiplexedStreams`. Unfortunately, using our implementation was
breaking Nailgun because it expected `System.out` and `System.err` to be
instances of `ThreadLocalPrintStream`. We were, however, replacing them
with another implementation, thus resulting in `ClassCastException`s in
Nailgun.

This commit removes `IdentifiedThread` (which was helping to distinguish
between the threads), and adapts `MultiplexedStreams` to use the
`ThreadLocalPrintStream`. `MultiplexedStreams` is kept to offer helpers
to initialize the streams and replace the streams to execute some code.

Fixes #159