Set the default nail class in case alias is incorrect
@olafurpg has correctly pointed out in
https://github.com/scalacenter/bloop/issues/173 that if the user doesn't
write the right alias, Nailgun will throw at it a weird message.

We fix this by setting the default nail class, so that if the alias or
fqn that the user types is incorrect we automatically redirect to our
entrypoint `Cli`.

This should finally fix all usability issues that Nailgun places on us.