Fix #207: Set status code OK before caching state
Otherwise, the subsequently loaded state will have a non-zero exit
status, which may "absorb" the success of the new operation. See for
instance the test `bloop.nailgun.BasicNailgunSpec.helpCommandTest` that
was failing. The reason is that we were putting in the cache the failure
caused by the previous test. This failure was then loaded when running
the `help` command, and the previous exit status was used.

Fixes #207