Cache source dependencies as git deps in CI
This patch allows the compilation of all these source dependencies in
the CI because it adds a dependency to a git commit hash instead of a
changing git submodule.

Sbt caches git commit hashes in the staging area and that area is
available for the CI across different runs, so it can be effectively
reused all the time.