Fix `refOf` for benchmarks
We use `refOf` to extract the commit reference out of our version
number, but the regex is no longer valid since we have a tag now.

We use this reference to know exactly with what commit to associate the
benchmark results and display the commit message next to the results.
Because extracting the commit reference didn't work, we couldn't find
the commit SHA and didn't upload the results.