Add warning about Bloop shell being unsupported
Incidentally, this fixes issues that we're having with the benchmarks:
We run Bloop shell in the benchmarks, and we're waiting to read '\n> '
to know that the prompt is ready for the next command. We used to log
'Loading xx projects from xyz...\n' when the shell started, but this
message has moved to debug. Therefore, because we don't run in debug in
the benchmarks,, we'd never read the first '\n', and the benchmark
would be stuck.