Remove `coursier` after building Homebrew formula
Because we download coursier in `install.py`, it was installed along
with the rest of Bloop in our Homebrew formula.

The easiest solution is to just remove it before copying the files. The
harder would be to have a build dependency on coursier, and use a
different installations script that wouldn't download coursier.