Implement a decent file watching task with monix
The previous file watching logic was a monstruosity because:

1. It relied on `Await.result` to get the state for every of the actions
   run on file events.
2. It was difficult to reason about.
3. It was not possible to cancel it.

This new implementation uses the nifty `Observer` and `Observable`
interfaces to implement file watching in a reliable way, closing the
watcher when the implementation is cancelled or whenever an exception
happens in the watcher.

This issue fixes https://github.com/scalacenter/bloop/issues/257, which
was caused because file watchers of previous clients (connections) were
still running in the background and acting on events.