Use a monix implementation that actually works
The previous one didn't work because `Observable` returns a list of
tasks, and that task cannot be processed and flattened until the stream
of file events has finished (never). This is bad, it means we cannot
fold left values produced by observables and then pass them in to
functions that return `Task`s.

For that, I've discovered you need `Consumer`, and there is one awesome
consumer that does exactly what we want: the one created with
`Consumer.foldLeftAsync`.