Propagate error to avoid downstream execution
As the compilation task does not throw an exception, but rather we model
failures in the DAG, we need to propagate the error down the task
dependency chain so that tasks that depend on compilation do not
continue if compilation has failed.

In this vein, this commit aligns the codebase to behave like this, and
then avoid testing a project that has not correctly compiled.