Use different classes directory than sbt
Fixes https://github.com/scalacenter/bloop/issues/344.

We don't do what the ticket describes yet because that's something left
for the future. So far, reusing the `.bloop` directory (which should
already be in .gitignore) for the target makes sense.

In the future, we'll get the sbt integration to talk to the bsp server
of bloop and compile to that directory, making the experience of
switching from sbt to bloop and back and forth smooth.