Update the build in every watch iteration
Nailgun shuts down the executor service running the "server code"
whenever an exception is catched. This usually takes around 1000s after
the execption has been thrown.

This exception is thrown after a SIGKILL or SIGINT, for example. The
problem is that in 1000s our bloop tasks do not have time to return, and
that means that the `Await.result` in `Interpreter` never finishes.

Therefore, the logic never reaches `Cli` back, and the state is not
saved. This code is a temporary workaround thought to overcome this
limitation. It fixes the current problem in which the compilations that
happen in file watching are thrown away. Now they are cached, and
subsequent compile commands won't trigger compile afresh.