Add flag to support source dependencies
The main issue with source deps in Bloop arises from the fact that, by
design, `bloopConfigDir` is build dependent, which means that it changes
across different source deps.

We don't want this to be the case for source dependencies though, since
Bloop needs all the project dependencies to be in the same directory.
For that reason, this flag aggregates the bloop config files to be
written to the same bloop configuration directory, and hence have it
work nicely with bloop.

Fixes #375.