Detect "compile->test" configurations in sbt
I detect this problem when using bloop. The configuration of the project
reference is important because a classpath dependency may point to an
arbitrary project dependency. An example of this is the "compile->test"
dependency to frontend from our `benchmark` project.

To solve this issue, we need to select the right name from the
configuration of the project ref, and only fallback to the configuration
of the scoped running task when that one is empty.