Use nailgun environment in forked processes
This commit propagates the `remoteEnv` properties in the nailgun context
to the place where we run and test projects in forked projects. Thanks
to the fact that we always fork, we're able to use the environment
variables from the user without weird hacks to mutate `System.getenv`
and scope it per client.

Fixes https://github.com/scalacenter/bloop/issues/335.