Use json as the configuration file format
This commit is quite big.

The reason it's so big is because in order to ditch java properties and
use the json configuration file format, I had to redesign some of the
parts of bloop's configuration handling that left me unsatisfied.

This includes ditching parallel collections to load the configuration
files, and reusing as much as possible the configuration classes within
the `Project` definition.

It also includes moving unnecessary loggers to the test project, and
removing how elapsed time and `verbose` loggers are handled (given the
use of `Task`, neither implementation was viable).

Lastly, it changes `Interpreter.execute` to use `Task`, so that we only
use `Await` once in the program: at the end, once we've captured the
control flow within a task.

Fixes https://github.com/scalacenter/bloop/issues/58.