Persist analysis files on the background
Previously, bloop would persist the analysis file on `shutDown`.
However, this hook has been removed upstream in
https://github.com/facebook/nailgun/pull/131.

As a result, we take another (more resilient) strategy to persist the
analysis file. We persist analysis file on the io pool after every
`Interpreter` execution, without blocking the return of the request.

All the logs are redirected to the nailgun server output.