Make results cache aware of previous compilations
(including the failed ones).

This is what this commit does:

1. `ResultsCache` no longer stores the successful previous results. Now
it also stores the `Compiler.Result`s, so that a state that has been
compiled allows us to derive summaries and now how the compilation was.
This includes being able to tell the user why a given project was not
compiled (because it was blocked by other ones). This is important to
return a `CompileReport` in bsp-land, but that's implemented in another
PR.

2. The `Result` ADT is moved to backend, `Success` now takes the
reporter too (to check how many error or warnings were outputted to the
user's log), and a reporter is derived from the source infos of previous
Zinc results.

3. The `Result` ADT no longer takes the project as a paramter, instead
we get the project from the `toCompileTask` in `Tasks.scala`.