Pass in framework-specific test arguments via CLI
For JUnit, for example: `bloop test frontend -- -z -a`.

Test arguments are difficult to pass in via the CLI because they require
a mapping to the framework. If they are passed in to the wrong
framework and the framework does not support the argument, it will fail.

In order to keep the ergonomics of the test CLI arguments to a sane
level, we don't ask the user to pass in the mapped framework (for now,
at least) because it's difficult to remember and they need to pass in
the fully qualified name of the framework they target. There is no way
users will remember that.

Instead, we try to cover the most common case: you only want to use the
test arguments in the CLI for codebases that only support one test
framework OR for commands that they specify a concrete test class via
`-o`. In those cases, we know for sure that there will be only one
framework involved, and then trust that the user passes in the right
test arguments to the command.

If this is not the case, we warn that the test options are ignored.

Remember that another way of modifying the test arguments is by adding
them to the json configuration file, but then they are removed whenver
they are generated again. This is not the best workflow to try arguments
out at the spot, hence this patch.