Report compilation logs via publishDiagnostics
The previous implementation was sending the compilation infos as if they
were normal logs. This patch splits up the reporters in two: one used
for normal logging and the other one used for bsp. The bsp reporter will
`textDocument/publishDiagnostics` all the compilation infos instead of
forwarding them with `window/showMessage`.