Add basic support for templates in install.py
Services in systemd do not use init files such as .bashrc, so we
cannot rely on bloop being in $PATH. Thus, we need to make
bloop.service a template and replace __BLOOP_INSTALLATION_TARGET__
with the path supplied by user as the argument to install.py.
This will also be helpful in case of XDG Desktop Entries.