Add freedesktop.org-compliant Desktop Entry
Adds a bloop.desktop file and an accompanying bloop.png icon
which can be used in freedesktop.org-compliant desktop environments as
either a simple shortcut or as an autostart item. The desktop entry is
made executable to let users run it without issues.
Updates install.py and user documentation as well.

The icon was generated with:
$ inkscape -z -e etc/xdg/bloop.png -w 256 -h 256 website/static/img/logo.svg