Don't overcompile failed children in seq compilation
When we compile sequentially, the users are entitled to pass in as many
targets as they want. These targets will then be reduced, and it can
happen that they are both root projects that require the compilation of
the same children.

Previous to that commit, bloop would try to compile failed children
several times. This error was discovered in the bsp implementation when
clients always pass in all the projects of the build graph, including
the root projects.

Now, we detect that we're doing sequential compilation and we fail
subsequent compilation of root projects if their children have already
failed compilation before.