Implement our Zinc frontend for pipelined compilation
Pipelined compilation works well across Scala codebases, but at the
moment Java code depends on Scala code, the Java compiler needs to wait
for the end of compilation all the Scala projects in that the Java code
could potentially depend on. This happens because Javac only knows about
class files.

Hopefully, this doesn't limit the speedup gained from pipelined
compilation much since Java code in every module of a Scala build is
rare.

However, whenever we want to implement that blocking wait, we have two
solutions:

1. Extend `AnalyzingJavaCompiler` ourselves and make it do a `.get` on a
   `CompletableFuture` from the outside.
2. Create our own Zinc frontend and encapsulate all the logic for
   incrementality with `Task`s.

The first one is a no-go performance-wise because we waste CPU cycles
unnecessarily: `.get` will hurt compiles that are happening at that same
time and that would slow down compilation more than we want it to.

The second one is our only option, at the cost of rewriting some of the
frontend logic of Zinc. However, this rewriting is also beneficial to
Bloop because it enables us to do the following tasks:

1. Provide better compilation logs enriched with Bloop's project information.
2. Have direct access to dependency analysis info and other compilation
   data that we can use later on to produce compilation reports and run
   statistics on them.

As a result, this commit syncs up with the latest Zinc version to
implement option 2. I must admit the code duplication is minimal, and
even more beautiful and well-organized than the original Zinc API.