Deduplicate code in link/run and add pipelined flag
The previous infrastructure was verbose, this commit reduces the
verbosity and makes the code more reusable by all the parts
(especifically the handling of errors between link and run).

The new changes also add `pipelined` to all the relevant commands (as
well as `incremental`, that was only present in the compile command) so
that users can enable pipelined from the CLI. If they do, then the
compilation will be pipelined.