Fix test errors and remove human readable format
The test errors have been uncovered by the new test event handler: we
forgot to pass in the test properties required for the tests to succeed.

Aside from fixing them, this commit removes the human readable format
because it's overly verbose and I find it difficult to read. Instead we
use a numeric version that has the metric on the rhs to be able to
quickly identify it.