Allow the project load of non-Scala projects
This commit allows Bloop to load up projects that don't have any Scala
setup. This can happen when build tools like Maven want to generate
configuration files for java-only projects.

This could have been workarounded by creating default scala options from
the build tool side, but that's uglier than just setting all the fields
of the `Scala` setup to empty.

This is a quick fix to unblock @lefou.

The ideal would be to modify the configuration file format to allow for
an optional scala setup, but that change would be quite breaking and
would conflict with some contributors' efforts that are already
undergoing, so I'll finish this off in a week or so.