Handle checksums better
Sbt doesn't seem to populate the checksum field in the resolved
artifacts. However, we know where checksums are typically resolved: in
the same directory as the target jar. In fact, it's almost the same path
except for the fact that the name of the algorithm is appended at the
end.

Therefore, we look for it and if there's we do our best effort to add it
to the bloop configuration file.

Let's remember how important this checksum is: language servers and
IntelliJ can use to detect when their indices are outdated and need to
be recomputed.