Add bloop keys for native and js config
I'm not sure the scalajs and scala native plugins define the information
of our configuration in the plugins. If that's the case, this approach
will require users to redefine these keys so that bloop knows about
their configuration.

If this is not good enough, we can discuss better and fancier approaches
to communicate with the cross-plugins without actually depending on them
as a binary dependency.