Beginning of mill integration (BloopModule)
Adds a `bloop.integrations.mill.BloopModule` mill external module that
can be used to generate bloop config from mill.

See https://www.lihaoyi.com/mill/page/modules.html#external-modules

The suppot is limited to JVM project at the moment, and does not fill
the "resolution" configuration element.

Usage :

1. add ```import $ivy.``ch.epfl.scala::mill-bloop:$bloopVersion` ``` in
the build.sc of a project (it's probably possible to add it to a
`~/.mill/ammonite/predef.sc` or a similar file instead)

2. run `mill bloop.integrations.mill.BloopModule/install` from the
directory of the mill project.