Clone and cache non-sbt integration projects
Add the sbt infrastructure required to clone all the required
integration projects that are not sbt and therefore are not handled by
sbt automatically whenever the project is referenced from the build
graph.

This approach will clone the integration projects at the beginning of
the sbt shell if they haven't been cloned. The logic to do so re-uses
the one that sbt provides us via `Resolvers.git`.