Unify pipelining and parallel compilation
The implementation now allows to run both at the same time, or let
users select which one should be applied for their projects and with
which configuration.

This unification also brings some nice cleanups and makes the whole
design much better to understand than before. Tests are left for a
future commit (soon).