Synchronize all java compilation
When running under pipelining, java compilation cannot be pipelined. For
a given module X that depends on Y, the java compilation of X cannot
start until the Y's java sources have been correctly compiled.

What this means is that projects that have a java->scala compilation
order will not benefit from java compilation at all, whereas those that
have mixed or scala->java compilation will be pipelined as much as
possible (synchronizing on all java compilations).

This PR implements the logic required for pipelined compilation to work
in the general case and prepares the pipelined community build for the
CI.