Add Finagle, Scalding, Algebird and Grid
... to our community build.

A comment on Finagle: Twitter uses Pants internally and therefore they
provide some tools (the checked-in dodo build tool) to publish snapshost
and be able to consume their projects from sbt. This projects adds the
required infrastructure to our build to set up twitter projects. In this
case, we start with Finagle.

Aside from that, we also:

* Cache .dodo products and run it with bash
* Don't use sftp for the CI cache (it's making the CI too slow)
* We upgrade to a new CI image that has openssl (required by dodo) and
  jdk11 for future testing.

We don't add Summingbird because of resolution errors (mergeErrors
failure in sbt, weird one, it's likely a bug deep down in sbt core).