Introduce `Dag.Aggregate`
`Dag.Aggregate` will greatly simplify the way we implement compile and
test when they take several projects. All we need to do is to aggregate
all the projects in an aggregate node and pass it to the current
implementations.

This commit also makes the implementation of `Dag.dfs` slightly more
efficient than before, assuming it will be a memory-intensive operation.