Sidestep project name collisions in sbt plugin
This issue was found thanks to algebird and scalatra, that make an
intensive use of `-` in their project IDs. This use unveiled a conflict
of names between a project `foo` and another project `foo-test`, where
the default name generated for `foo-test` would collide with the name
generated for the test configuration of the project `foo`.

This had the effect of writing to the same file twice, and writing the
values of a dependency that caused a cycle in the DAG. This violation
was then catched in Bloop's DAG checker.

Now, we make our best effort to pick a name that doesn't collide with
the existing one.

This commit bumps up the `BuildPlugin` version to force the regeneration
of bloop configuration files for algebird and scalatra.