Initialize internal exit status list after sourcecode
Fix an issue where bloop would run `test` or `run` even if `compile`
failed. This issue happened because of a bad initialization error in the
`ExitStatus` object that was causing `codeToName` not to work. This
affected the following logic in `Interpreter`:

```
runCompile(cmd, state, project, deduplicateFailures, excludeRoot).flatMap { compiled =>
  if (compiled.status != ExitStatus.CompilationError) next(compiled)
  else Task.now(state.withDebug(s"Failed compilation for $project. Skipping $nextAction..."))
}
```

After the change to the exit status, the test passes.