Don't fail in builds of single sbt plugins
I knew about the possibility of failing in builds of single sbt plugins,
but however didn't protect us about them because I hoped sbt would
expose an entrypoint to know for real if a project is a meta build or
not. As this hasn't been the case, I've implemented a simple heuristic
to know whether a build is a meta-project or not: check if the CWD is
the same as the root of the build.