Replace `debugInContext` by `debug`
The solution required to do this change without modifying
sbt.testing.Logger's interface relies on the linearization order and the
check for ambiguous overloads. Typically, if two methods are defined in
an interface and they have the same public interface, except that one
takes an implicit and the other one doesn't, the compiler fails with an
ambiguous overload error when one of the methods is used (e.g.
`logger.debug(msg)` refers to `logger.debug(msg)(implicit ctx)` or
`logger.debug(msg)`?).

To fix this problem, we override the version that doesn't take a
parameter in a super class of `BloopLogger`, which is the common logging
interface across bloop and is implemented by many different types of
bloop loggers. Then, we add two abstract methods that must be
implemented by subclasses:

1. An implementation of a `private[logging] def printDebug`, which is
   used by the overridden simple `debug` in `BaseSbtLogger`.
2. An implementation of the long debug method that we want the compiler
   to select for every use of `logger.debug(msg)`.

This trick helps us remove `debugInContext` and simplify the logger API.
From now on, everyone will be forced to use `debug` and provide a
context if there is not an implicit one.