Fixes cancellation in Scala.js 1.x
Scala.js 1.x's `ComRun` doesn't invoke `close` on the underlying run
when its `close()` method is called. This means that the process is
effectively not killed and a resource leak happens. This leak can happen
fatal consequences in Windows because an open reference to the JS output
file is open and, therefore, other processes cannot write to it and
fail.

To fix this issue, we've had to copy the contents of `BloopComRun`,
`JsDomNodeJsEnv` and `NodeJsEnv`. These environments use nuprocess to
spawn the JS test servers.

When Scala.js fixes the aforementioned issue upstream, we can remove all
of this logic and keep the js bridge 1.x logic as simple as that of
0.6.x. The use of Scala sys.process doesn't bother us because
cancellation still works successfully.