Add partial migration to bsp 2.0
It's a partial migration because we have not yet adapted to bsp 2.0 for
test, run and a few other endpoints. However, we implement the most
urgent endpoints required for BSP clients to start their migration to
2.0.

The way we test compilation in the tests has changed substantially. In
this commit, we introduce a fine-grained way of testing diagnostics as
well as other compilation invariants.

In general, this commit adds:

1. A clearer separation of diagnostics reporting for CLI and for BSP.
   (Both have opposing interests when it comes to reporting diagnostics)
2. In-depth tests that check the correct behaviour of BSP reporting
   for clients, including reporting whenever errors are cleared or
   whenever there is a no-op incremental cycle.
3. A better way of getting previous compilation problems where we don't
   depend on the source infos of the Zinc analysis.