Add resources to the configuration file
Resources are files that are used at runtime, either when testing or
running an application. These files can be generated by the build tool
or present in a resource directory of a project.

Resource files have been supported in bloop since 1.0, where in sbt and
other build tools we were relying on the resources to be copied by the
build tool to the classes directory. This approach had a problem:
whenever a change is done in an unmanaged resource file (a file that is
controlled by the user, like a `log4j2.properties` file), bloop would
not pick it up until there's a new `bloopInstall` invocation.

This is a known limitation that this commit fixes. The approach to
support resources has changed:

1. A new `resources` field has been added to the JSON file.
2. Resources are not copied to the class file, they are instead added to
   the test and run classpaths.

The second item makes a subtle but important point. The fact that we
don't copy the resources means two things:

1. We don't care about the relative path with regards to the root
   resource directory where a resource lives.
2. We rely on the contents of the `resources` field, which should
   contain only those directories which are either considered roots (e.g.
   files from where the application can use relative paths to look for
   resources) or single files.

We support this treatment of resources in all build tools. Sbt requires
a special case in the plugin because the source of truth is `resources`,
which contains all the contained files in the resource directories + any
other file the user may have added. As a consequence, we filter the
paths until we get either root resource directories or single source
files. Note that removing directories that are already contained in a
root resource directory is really important because it dictates the way
the resource will be accessible at runtime (where the lookup relative
paths play an important role).

Fixes https://github.com/scalacenter/bloop/issues/586