Handle build load errors gracefully in the server
Imagine that `bloopInstall` fails to successfully complete. The process
did complete for a subset of the build graph, but not all of it. And now
imagine that the errors are not relevant and the user is willing to
ignore them temporarily. Isn't it frustrating that bloop will fail to do
anything with a `NoSuchElementException` because the build load fails?

This commit tries to improve on this situation by not making it an
all-or-nothing situation. Bloop can now handle missing dependencies in a
configuration file gracefully. To let the user know that there's
something off, though, we emit a warning every time we compile a module
that has a missing dependency (in case, for example, there's a failure
in compilation). However, we do not emit such warning if the user
compiles other modules in the build and therefore give more freedom to
the user.

Fixes https://github.com/scalacenter/bloop/issues/708