Use another hash code for projects
The previous `hashCode` of `Project` was using the hash of the
serialized file. At the time of doing that, I did not realize the use
case that this commit fixes would be affected.

We're now using a unique ID composed of the name and the configuration
path a file comes from. Whenever that is different, it must be a
different project.

We use `ByteHasher` instead of `hashCode` to avoid hash code collisions.
The probability of a hash code collision is very slow, especially
because the server should only have around 100 or 200 projects maximum,
but it's better be safe than sorry.

Fixes https://github.com/scalacenter/bloop/issues/733