Add mechanism to force test resources generation
We seem to get some spurious errors in CI after some changes in latest
master. These errors look related to caching the test resources.

With this commit, the user can add `FORCE_TEST_RESOURCES_GENERATION` to
the commit message and the CI will force the generation of the test
resources.