Add `bloopMainClass` to sbt plugin
In sbt, we cannot reuse `mainClass` because it triggers compilation when
the user has not specified it in its build.sbt. From the outside, there
is no way for us to know if the user declared it or not, and therefore
we plainly cannot trigger the execution of `mainClass`.

To allow users to specify the main class though, we introduce a
`bloopMainClass`, a duplicate of `mainClass` which doesn't trigger
compilation, so that users that don't want to verbosely pass in the main
class every time their application can set it in their build.sbt file.