Fallback to direct server invocation in nailgun client
The python nailgun client allows to run the bloop server when it's an
executable (a script) and not a jar so that running `bloop server` in
systems such as Nix works and doesn't fail with an error.

This mechanism is only a fallback iff the invocation with `java -jar`
failed.