Add `--skip-bsp-connection` mode
This mode is important for clients that want an automatic way of
installing bloop and spawning a bloop server in the background, without
the requirement of opening a bsp connection.

The launcher can be something they can solely depend on and execute in
process or shell out to if they resolve it with coursier and have a path
to it.