Reorganize tests in frontend
The organization of test sources in frontend was chaotic and there was
no coherence in the way tests were defined. This new change moves around
a lot of code to bring some coherence to the test sources in the
codebase and make it easier to see what things are tested and not. We
follow the pattern `feature name + Spec` for every feature that we test.

This commit also makes some small refactoring to the code to polish some
rough edges of the API (such as the definition of a `bloop.monix`
package that would prevent the use of `monix` APIs without prepending
them with _root_).