Clear previous problems as early as possible
BSP requires that bloop clears problems that were present in previous
compiler iterations manually. Previously, Bloop would meet this contract
by clearing problems from previous compiler runs at the end of a
successful incremental cycle. For every incremental cycle, bloop would
go through the problems and clear those that were affected by changes
occurred during that cycle. You can see more information in this pull
request: https://github.com/scalacenter/bloop/pull/782

However, this approach was slower than desired and required that we
successfully finished an incremental compiler cycle before *starting* to
clean some diagnostics in the recompiled files.

Instead, this pull request takes a different approach to this problem.
We clear diagnostics as soon as we know they are safe to be cleared. For
example, if in a previous compiler run typer emitted an error, we have
bloop record that a diagnostic happened during that phase and, in
subsequent compiler runs, we clear the diagnostic as soon as we've
passed typer for the source file that contained the error.

That is, we clear diagnostics right after the phase that generated them
succeeded.

If a diagnostic could not be mapped to a phase, we fallback to the
previous behavior: we will clear it in `reportRemainingProblems` right
after every incremental cycle.