Include resources in dependency classpath
In https://github.com/scalacenter/bloop/pull/730, I claimed:

> Resources are files that are used at runtime, either when testing or
running an application. These files can be generated by the build tool
or be present in a resource directory of a project.

However, this was not the full story, it turns out that resources are
also required by macros, which happen at compile time, and therefore
resources are required by `compile`.

This patch modifies bloop to use the classpath with resources as the
default from now on for any action in bloop, as all of them require it.

Fixes https://github.com/scalacenter/bloop/issues/763