Reduce footprint of cached state
Before this commit, the state cache used to cache all the fields of the
state, including those that were client-specific or part of a concrete
bloop execution (such as exit codes).

This commit introduces an internal abstraction `CachedState` which only
represents the core of the client state, which is commonly referred as
just `State`. This core state is only visible to the state cache and,
upon every query, the use site passes all the client-specific stuff
needed to create a `State` from a `CachedState`.

This change is small because it hardly leaks to the rest of the bloop
APIs (it was designed with such a property in the past).