Use a guaranteed-unique name for temporary directories
This is an attempt to correct race-conditions relating to concurrent
access of a temporary directory called `classes.bak`. The change chooses
a different, unique name (a UUID) for the temporary directory.

Bloop's build is currently failing, so I can't test this change, but
it's very simple. But I'm relying on a review, or some evidence of
passing tests, to confirm that this is the right thing to do.