Rework compile IO and optimize no-ops
1. Implement a fast classpath hasher.
1. Implement a fast source hasher.
1. Add more tracing to the compilation engine.
1. Hash classpath and sources via xxHash. Adapt that xxHash to an
   hexadecimal representation so that it can integrate with Zinc APIs.
1. Reuse source hashes and classpath hashes in the detection of
   `InitialChanges` in `BloopNameHashing`.
1. Don't compute the bundles upfront. For a given parent node in the
   project DAG, the bundle must be computed only when the children have
   been compiled successfully.

All changes have been done after extensive measuring in a sluggish iMac
with few memory, as it's more or less representative of what normal
Scala developers will use to compile their code. In this kind of
machine, IO work is more expensive.

The commit also includes some changes to do copying of class files after
compilation to an independent classes directory. For now, it's been
disabled and will be enabled in a future commit when we make the full
conversion to the new read-only incremental compilation scheme.